pyproject.toml
[project]
name = "customersupportagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-adk>=1.19.0",
    "google-genai>=1.52.0",
    "google-generativeai>=0.8.5",
    "greenlet>=3.2.4",
]
